Move video player to be part of the util module
