+++ src/util/media/src/androidMain/kotlin/com/gabrielbmoro/moviedb/media/VideoPlayer.android.kt
@@ -1,4 +1,4 @@
-package com.gabrielbmoro.moviedb.details.ui.widgets
+package com.gabrielbmoro.moviedb.media
 
 import android.view.ViewGroup
 import androidx.compose.runtime.Composable
