+++ src/util/media/src/iosMain/kotlin/com/gabrielbmoro/moviedb/media/VideoPlayer.ios.kt
@@ -1,4 +1,4 @@
-package com.gabrielbmoro.moviedb.details.ui.widgets
+package com.gabrielbmoro.moviedb.media
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
